fix: use init that makes containers handle SIGTERM
This change requires Docker engine 18.06.0+ to be installed.
That's more than 2 years ago so I think that's ok for our modern dev.
`electric` and `vaxine` containers don't seem to handle SIGTERMs
in how they implement entrypoints, so this makes them behave.

A more universal approach would be to bake in `tini` or smth like
that into the docker images, but this is faster for now.

References:
1. https://vsupalov.com/docker-compose-stop-slow/
2. https://github.com/krallin/tini