[VAX-532] Introduce dependency info for row add-wins semantic
Tags mechanism is added to Satellite protocol to track row
dependencies in the form similar to ORSet, to provide add-wins
semantics for concurrent operations. Tags stored with the row
determine wherever or not row exists. Previously row existence
would be determined based on the crdt flag, this was changed to
a set, containing tags.

Commit also removes REPLICA IDENTITY DEFAULT support, as it
was identified, this will not work with this approach.

Implementation brings non-comparible change to Satellite protocol
 and requires Satellite protocol version 1.0