ci: Disable LCOV reports in pull requests
Here's an example of such report -
https://github.com/electric-sql/electric/pull/100#issuecomment-1439689740.

These reports add a lot of noise to long-running PRs. BTW, the comment
linked to above seems to have been edited by the github-actions bot to
keep it up-to-date with the latest CI run for that PR. So perhaps
there's a configuration settings that disables posting of additional
comments with LCOV reports within the same PR?

Regardless, we don't have a defined target for maintaining code coverage
and maintaining unit test coverage is currently not a priority while
we're focused on rearchitecting Electric and working towards the v2
release. Removing these reports will also make the open-source repo a
bit more friendly to new contributors since they'll have one less thing
to worry about when submitting their first PR.