fix: insert all column values on update (#139)
Fixes VAX-659: Updates are translated to an insert of the modified
columns. However, if some columns are not updated, they will not be part
of the insert query and thus we lose those values.

This PR fixes this issue by using the complete row in insert. However, we need to be careful that this row contains the latest information for all columns, i.e. the incoming entry may have a bigger timestamp than the local change but if some column was only changed locally we should take that value and only use LWW for columns that were modified by both the incoming and the local change.