Remove all uses of ConsoleClient and related code
The only use case for ConsoleClient was to request an auth token from
the console service. We're now rebuilding our approach to auth and getting rid of the
dependency on console is part of that.

Now that the TypeScript client does not depend on the Electric console
for obtaining test tokens, we have to provide an explicit one in
e2e/satellite_client.