feat(satellite): build migrations on Satellite (#181)
This PR builds the migrations that are fetched from the Electric
endpoint.
The built migrations (i.e. DDL + triggers) are written to the Electric
configuration file: `.electric/@config/index.mjs`. This file is loaded
by Satellite at startup in order to apply new migrations.

In a nutshell, this PR does 3 things:
- read meta data for the fetched migrations
- generate triggers for the migrations
- write the migration and its triggers to `.electric/@config/index.mjs`

The main bulk of work is in `clients/typescript/migrators/builder.ts`