Add satellite collector process pair
This is required to "feed" Postgres transactions from all Satellites
at the same time. Satellite WS server need a consumer to initiate
the subscription, so we're giving it a consumer which in turn
passes the transactions on to the producer. Producer stores
these transactions in ETS, and can then act as a producer for
the Postgres SlotServer.