feat(ts): added genUUID function (#184)
As discussed in https://github.com/electric-sql/examples/pull/22

It seems that, even though the Crypto API is not generally enabled in
insecure contexts, the `crypto.getRandomValues` function is
[exempt](https://developer.mozilla.org/en-US/docs/Web/API/Crypto/getRandomValues)
from this, unlike `crypto.randomUUID`, which means that generally, it
should also be available to our new `genUUID` function. There is a
fallback to `Math.random()`, just in case.

I made a new function because the `randomValue` function was already
used and the format is different from `genUUID`. It may be overly
cautious though.

A bit on the original problem with concurrent calls: using
`Math.random()` alone or just taking its output and deterministically
transforming it into an uuid suffers from the same problem. From the
documentation, it seems that `crypto.getRandomValues` should side-step
this problem.