feat(cli): npm migration script (#201)
This PR introduces a NodeJS migration script that fetches all migrations
from Electric and uses them to upgrade the application. It is written as
a NodeJS program to ensure cross-platform compatibility. The
`package.json` is extended with a `bin` property allowing users to write
`npx electric <command>`. At the moment, we support a single command
(`migrate`), so `npx electric migrate`. When calling `npx electric` it
invokes the NodeJS script, and that script than looks at the additional
parameter `migrate` and invokes the right handler for that. So in the
future, we can extend this CLI with support for other commands.