feat!(pg): Stream Postgres replication log to Satellites
This commit implements direct streaming of changes from Postgres to
Satellites bypassing Vaxine.

It's implemented by utilizing a special WAL cache: a process, that
stores WAL segments coming from PG, and the Satellites know and track
their position in this WAL cache. This is because Postgres really
doesn't like a lot of logical readers, so we have just one, and do the
fan-out from the cache. Clients may also be in different positions in
the stream, which leads to the WAL Cache interface providing options to
start from an arbitrary point.

The ETS-backed WAL cache implementation is a bit rough, but it's to
get it going at all. We're likely to replace it with an improved
implementation a bit later (there's a PR by defnull that implements the
improvements, but I decided to merge that PR after we have a baseline
working

The LSN position that Satellite stores is opaque, however with the
coming shapes/initial sync logic, client may need to be able to compare
those LSNs. For this reason it's a string-serialized number instead of
`:erlang.term_to_binary/1`.