feat!(pg): move conflict resolution to PG using special triggers (#206)
This commit addresses the "write path" from Satellites to PG.

It introduces special conflict resolution triggers, that are installed
and/or updated for electrified tables. The general logic is that each
Satellite operation is mapped to two operations on PG: one on the shadow
table, and one on the main table, with main table operation coming first
in the replication stream to keep with the lock order that PG does on
transactions that are done via psql. When receiving the writes from PG,
we also need to "merge" shadow and main table operations (although we're
interested only in one column, `_tags`, from the shadow table. We also
use the `_tag` column as a transient value - it doesn't get
conflict-resolved, which means that for every transaction it'll stay the
same. We're using that knowledge to restore timestamp/origin of the
transaction.

## `electric.tag` type

This is a special timestamp, which is a pair of a wall-clock timestamp &
origin. We're ordering operations by that timestamp, with `NULL` origin
on Postgres ordering higher in case of equal timestamps - this is to
make sure PG always wins conflicts to keep expected PG write semantics.

## Shadow table structure

Given a table:
```sql
CREATE TABLE public.entries (
    id uuid DEFAULT public.uuid_generate_v4() NOT NULL,
    content character varying(64) NOT NULL,
    content_b character varying(64)
);
```
We expect a shadow table structure:
```sql
CREATE TABLE electric.shadow__public__entries (
    _tags electric.tag[] DEFAULT ARRAY[]::electric.tag[],
    _last_modified bigint,
    _is_a_delete_operation boolean DEFAULT false,
    _tag electric.tag,
    _observed_tags electric.tag[],
    _modified_columns_bit_mask boolean[],
    _resolved boolean,
    _currently_reordering boolean,
    __reordered_content character varying(64),
    __reordered_content_b character varying(64),
    id uuid NOT NULL,
    _tag_content electric.tag,
    _tag_content_b electric.tag
);
```

Information about how the conflict resolution triggers work will be
available in https://github.com/electric-sql/pg_triggers.

## Collector processes

We need to merge all transactions coming from Satellite instances into a
single stream feeding PostgreSQL, while also having a GenStage
subscriber to the `WsServer` process to initiate streaming from
Satellite to Electric. This is solved by a pair of processes:
`SatelliteCollectorConsumer` and `SatelliteCollectorProducer`. The
producer here acts as a GenStage producer to feed `SlotServer`, and gets
the data from an internal ETS table, acting as cache in case Postgres
needs to reconnect. The consumer generates demand for the websocket
process, and stores the events sent by Satellite forwards the events to
the producer to be stored in the ETS and eventually fed to PostgreSQL.