feat(CLI): watch flag for generate command (#217)
This PR introduces an optional `--watch` flag for the CLI's `generate` command. When running with the `--watch` flag, the process periodically polls Electric's migration endpoint for new migrations. The flag accepts an optional argument which is the polling interval in milliseconds (i.e. how often to poll the endpoint).
Use it as follows:
```sh
npx electric-sql generate --watch [<pollingInterval>]
```

We modified the Electric http endpoint to return a 204 if there are no new migrations and added tests for that.