Keep track of the last seen migration xid during initial sync
When replicating migrations to the client during initial sync, we may
see the same migration coming from CachedWal to due the lag inherent in
logical replication. To avoid sending the the same migration to the
client more than once, we store the last known migration transaction xid
in WsServer process state and check subsequent migration transactions
against it.