fix(electric): tag serialization wasn't working assigning `null` to tags (#230)
Internal representation of tags on PG is done in a way that any
PG-originating tag has it's origin component set to NULL, and then NULL
origins are ordered higher than any strings. When PG-originating tags
are sent to Satellites, we preprocess tags to replace NULL values with
pg origin name. But the inverse path has an error, whereby on a return
trip the pg origin in observed tags is not set back to NULL before being
set to PG like it should.

Closes VAX-762