feat(electric)!: implement partial replication via subscriptions (#222)
This implements partial replication using subscription. Client-side part
of this has been done in #196, so this implements the server-side. This
implementation is a breaking change, as clients that do not issue
subscriptions will never receive any data, unlike previously.

## Subscription logic

When a subscription request comes in, we need to do two things to send
the initial snapshot: request the data and figure out where to insert it
in the replication stream, especially with other subscriptions already
active.

We achieve the correct functionality with two techniques: `REPEATABLE
READ` transactions and stalling the outgoing replication stream if we've
reached the point of insertion before we actually have the data for it.

### Insertion point

Under `REPEATABLE READ` read-only transactions, we can confidently say,
that
[`xmin`](https://www.postgresql.org/docs/current/functions-info.html#FUNCTIONS-PG-SNAPSHOT-PARTS)
of the current snapshot (i.e. the last transaction id that has been
committed and visible) is definitely the point after which the have not
seen any transactions. That means that while processing the replication
stream, if we know the `xmin` of a subscription initial data request is
`X`, then we can safely send any transactions where `id < X`.

### Pause logic

When subscription initial data is requested, we start the aforementioned
`REPEATABLE READ` transaction from a separate process and wait for
`xmin` to be relayed back to the main process. This means the WebSocket
process if frozen until we know the insertion point (which should be
very fast) to avoid accidentally skipping past the insertion point.
There could be a case where the client send multiple subscription
requests while the first one's data is still pending. In that case we
use the knowledge that the `xmin` of the next subscription request is
going to be larger or equal to the previous one, so we enqueue the new
insertion point to be handled afterwards.

After the insertion point is known, there are four possible scenarios:

1. We've been sending the replication stream, but we've reached a point
where transaction id >= `xmin` for the given subscription. That means we
cannot continue until we are able to send the subscription data. We thus
pause the outgoing stream and enqueue all events that are meant to be
sent and wait for the initial data to arrive. Any additional events that
come in while the stream is pause are also enqueued. We may already have
the data for the pending subscription (see scenario 4), in which case we
can send it immediately and continue sending the events until the next
insertion point.
2. We've received the subscription data while the stream is "paused",
and we're paused waiting for this subscription. In that case we're
immediately sending the data over to the client and then continuing
sending the events, starting from the enqueued ones, until the next
insertion point.
3. We've received the subscription data while the stream is "paused",
and we're paused at an insertion point for a different subscription. In
that case we just store the data to be sent when we reach the correct
insertion point.
4. We've received the subscription data while the stream is ongoing.
That means we've definitely not reached the insertion point for this
subscription, so we store the data to be sent when we reach the correct
insertion point.

### Filtering logic

The entire "insertion point" logic exists to support correct filtering
of the outgoing transactions: we should send any operation that aren't
part of any subscriptions. That rule works with initial data seeing only
transactions with ids < `xmin`: any changes pertaining to the
subscription before `xmin` will be filtered out, and after we send the
initial data "at" `xmin`, any changes after will be sent through and the
client will see the correct picture - both initial state and the
streaming changes. It means that only subscriptions for which the data
has been sent are considered "established" and are used to check if the
changes fall into one of the subscriptions.