feat(electric): support unsubscribe messages (#234)
Adds support for the `SatUnsubsReq` message, that stops ongoing
subscriptions. That means the client will no longer receive data for
that particular subscription, and any yet-unfulfilled subscription
listed in the unsubscribe request will get immediately cancelled and
their data discarded.

Note: E2E tests are missing because although `unsubscribe` call can be
issued by the client, we're yet to expose a public interface for it to
e2e test.