Fix e2e flake by avoiding matching on normal output and log output at the same time
In Elixir, logs can be printed out of order with the regular output.
For example, having these two lines in the code

    IO.puts("hello")
    Logger.debug("world")

Does not guarantee that "hello" will be printed before "world". On the
other hand, multiple log entries are guaranteed to be printed in the
same order in which they were logged in the code, because a single
logger process receives all log entries and processes them in sequence.

This change replacing the match on the prompt with a match on a log line
that has the same effect for the purposes of the
`client_session_continue` macro here.