fix: make sure the subscribe call is held off until the replication is established (#271)
We've checked if the replication is already established in the
`subscribe` call, but we actually don't wait (or even expose a way to
know) if the replication has been established. This addresses one part
of it: the `subscribe` call will wait until `startReplication()` call
resolves after `connect()`. That way we can reliably send the
subscription request only after the replication is started, even if
`.sync()` call is made immediately after `await electrify(...)`. Any
fails in `connect()` or `authenticate()` will abort the `.sync()` call
with the same error, since at that point we cannot do any online calls.

We may still want to the result of `startReplication()` on `electrify`
call, but that's for another PR.