fix(generator): solve casing issues in table names (#302)
This PR solves the casing issues with table names.
For snake_cased table names we convert it to PascalCase and use that as
the model name.
If the table name is not snake_cased but starts with a lowercase then we
convert the first letter to uppercase and use that as the model name.

Note: the model name is only used for generating the types of the
Electric client as well as the Zod validation schemas.
However, the generated `tableSchemas` object still uses the actual table
names such that programmers continue to access the tables using their
original name, e.g. `electric.db.user`.