client: fix buggy connectivity hook. (#298)
The problem was that existing satellite processes were not reporting
their existing connectivity status to the connectivity hook. This was
confusing the hook into always starting in a disconnected state, despite
the connection being connected.

I've left one note in the code about a future refactor. I'm not
convinced by the current code structure and connectivity state machine.

---------

Co-authored-by: Ilia Borovitinov <ilia@electric-sql.com>