chore(starter): improved output and commands (#332)
This PR hides output from the underlying processes when running `npx
create-electric-app my-app` as well as in the output of the yarn
commands. Also introduces a new command `db:connect` that connects to
the PG database using the `DATABASE_URL` env variable. If that variable
is not defined it assumes that we're running the entire stack and it
will connect to the PG that is running in Docker.

---------

Co-authored-by: Valter Balegas <balegas@electric-sql.com>