fix: make sure that if slot server dies, it can continue from a known point (#375)
Original issue is that when PG cuts the TCP connection that is sourcing
the replication from Electric, the process structure wasn't implemented
in a way that allowed correct continuation of the replication stream to
PG without losing operations by the clients.

1. I've moved in `OffsetStorage.save_pg_position` into the loop because
the loop crashes on a closed connection when trying to call the `send()`
function that matches on `:ok`
2. When PG reconnects, it will tell us the latest LSN it persisted - we
use the saved mapping to restore that position in the cache and now we
can continue from there.
3. If all `send()` calls have succeeded we notify our producer that the
Satellite-sent txs can be garbage collected

There are still two issues after this PR:
1. We're calling `ack` on the Satellite txs within this producer, which
is a problem on the server restart, as they've not yet been persisted.
Proper fix would be to listen for a ping message from PG and ACK only
then, but there are yet more issues with that approach.
2. If the a "bad" txn is last in the batch, the PG will close the
connection, but only after the last `send()` call, so we will have
called garbage collection too early and may lose a write. This requires
a more systematic fix.