fix: run GC before applying the incoming txn
Problem was that GC was called after the application, so when a
transaction did a round-trip to the server and back, it would then
be included in the conflict resolution against itself, which is
a big problem if the tx was altered on the server-side conflict
resolution process. Removing these transactions first ensures
we don't have this problem.