Remove client's remote_ip from the default Logger metadata
IP addresses are user-identifiable information protected under GDPR. Our
customers might not like it when they use client IP addresses in the
logs of their on-premises installation of Electric.

Although it appears the consensus is thta logging IP addresses is fine
(see https://law.stackexchange.com/a/28609), there are caveats.

I think that adding IP addresses to logs should be made as part of the
same decision that determines the log retention policy. Since we're not
tying the logged IP addresses to users' personal information managed by
customer apps, we cannot clean them up as part of the "delete all user
data" procedure that app developers have in place to conform to GDPR
requirements. Therefore, logging IP addresses by default is better
avoided in production builds of Electric.

We may introduce it as a configurable option for better DX later on.