chore(electric): Clean up the response returned from /api/status (#412)
We should review the usage of connectors in the code base and refactor
those bits of code now that we have a single connector. In this PR I'm
only addressing the issue raised on Linear because it made sense to me
to make the Status API return a human-readable explanation of the
server's status.