Relax the client-side migration version validation
The server does not impose any restrictions on the migration version
string, although it implicitly assumes versions' lexicographic order to
match the chronological order of their application.

This change does not affect the ordering of migrations but gives
developers freedom to use a descriptive suffix in their migration
version strings.