fix (satellite): avoid applying migrations twice (#516)
This PR fixes the bug where a migration would be applied twice in the
following scenario:
- migrate PG while client app is running
- re-generate client
- restart app

On restart the app would try to re-apply the migration it has previously
received through the replication stream.

Normally this should not occur because we keep track of which migrations
have been applied. However, we forgot to store the migration version in
the transaction (on Satellite) which caused it to be processed as a
regular transaction instead of a migration ([cf. this if test in the
code](https://github.com/electric-sql/electric/blob/main/clients/typescript/src/satellite/process.ts#L1230))
and hence we did not mark it as applied.