fix(starter): [VAX-1187] Modify db:psql to connect to PG via Docker container (#538)
I've defined a utility function to run `docker compose` without shelljs,
and used that for both `backend/startCompose.js` and `db/connect.js`.

The benefit is that we get output coloring back.

The drawback is inability to capture the output and inspect it for error
messages. But the current code that's supposed to find the port reuse
error in the output already does not work as expected on Linux because
the error message printed there is `address already in use`.