Do not create an operator class for electric.tag
Only a superuser can create an operator class but we're moving away from
the superuser requirement for Electric.

An operator class is used in the following situations:

  * an index is created on a column that has a user-defined type with
    custom operators defined for it

  * values of the user-defined type appear in the ORDER BY clause

  * values of the user-defined type are used in DISTINCT or GROUP BY
    clauses

I haven't been able to conduct an exhaustive audit of the code to
confirm that we don't use the electric.tag type in those contexts. So
I'm proposing we merge this change into main and see if that causes any
issues in how Electric works.