feat(DAL): support for uuid (#547)
This PR adds support for UUIDs to the DAL.

Summary of changes:
- The generation script (`yarn client:generate`) was slightly modified
to add a `/// @zod.string.uuid()` comment to the introspected Prisma
schema. This comment instructs the Zod generator to extend the Zod
schema with a uuid validator.
- The generator is changed to expose the UUID type in the generated
Electric client
- An e2e test was added to test that UUIDs can sync to and from Electric
and that writes with invalid UUIDs are rejected by Zod.