Remove speculation about username and dbname as applied to the Electric proxy from the docs
As we're getting to ready to move away from relying on a single
"superuser" role for all Electric's connections to Postgres, it's
becoming increasingly likely that the username passed to Electric Proxy
will be used with upstream connections to PG as well. In other words,
the Proxy will have to start honoring the provided username and switch
its operating mode depending on the database name instead.