Switch from "electric" to "postgres" as the default DB user
The username "electric" is problematic due to the confluence of these
two things:

  * Postgres' default search path is `"$user", public`.
  * Electric creates and maintains a schema named `electric`.

With those defaults, when a user executes an SQL statement such as
`CREATE TABLE` and doesn't qualify the table name with the schema name,
e.g.

    CREATE TABLE items (...);

the result is that the `items` table is created in the `electric`
schema.

The username "postgres", on the other hand, is the default "superuser"
role and doesn't suffer from schema name collisions.