chore(examples): Update Linearlite to use the Proxy and new types (#590)
This updates linearlite to use the Proxy and new types.

Of the new times we are using `timestamp` and 'uuid'.

The only difference between the files from the starter and their copies
in this example is that they have been renamed from `*.js` to `*.cjs`.
This example has `"type": "module"` set in `package.json`, by renaming
the scripts to `*.cjs` they continue to work unmodified. Without this
they would have to be converted to ESM with "import" rather than
"require" syntax.

`builder.cjs` is included in the project but not used to build the app,
it uses Vite. It is inspected for port definitions by other scripts from
the starter and so cannot be deleted without modifying them.