feat(examples): New Ionic demo app - Appointments (#499)
This is a demo combining Electric with Ionic - a simple
Appointment/Booking app.

The Ionic CLI was used to start an Ionic app, I then manually added the
scaffolding from our starter. This process can be seen in the first two
commits.

It now uses the latest Electric with the proxy, and exercises the new
`uuid`, `timestamp` and `bool` type support.