fix(client): modify expo-sqlite driver to work on Android (#592)
This PR addresses #473 for expo applications running on Android.
The crux of the problem was that we execute SQL queries containing a
`RETURNING` clause but the sqlite driver for Android did not support
this. To address this issue, we use the new `execRawQuery` which is
provided by expo-sqlite as a workaround (cf.
[https://github.com/expo/expo/pull/24320](https://github.com/expo/expo/pull/24320)).
However, this requires rewriting our database adapter for expo-sqlite
such that it uses `execRawQuery` instead of expo-sqlite's transactional
support. As a result, we now have to manage transactions manually using
`BEGIN` and `COMMIT`/`ABORT` statements. This is very similar to the way
how our database adapter for wa-sqlite works. Therefore, i decided to
extract that implementation into an abstract `DatabaseAdapter` class
(cf. `clients/typescript/src/drivers/generic/database.ts`) and have the
expo-sqlite and wa-sqlite adapters extend it.