fix (DAL): transform all data types in input structures (#634)
This PR fixes the bug reported by Kyle Mathews on Discord:
https://discord.com/channels/933657521581858818/1079688869852753981/1169710877201674300

> Possible bug in 0.7. Upgraded and changed column type to boolean now
my node.js liveMany query is failing with SQLite error that it can't
bind to the type past in. In the better-sqlite adapter, if I log out
queries and args passed to the query function then I see a "true" is
being passed in which presumably should be 1. If I try to change the
query back to use 1, I get a zod error (as I'd expect).

The problem came from the fact that we were only transforming date
objects to SQLite strings but not the other values that were added later
(booleans, floats, etc.).