Change useState to useRef for values not used for rendering
Thisis the general rule for state vs. ref — only use state if it's
rendered — which in the case of useLiveQuery is just the results.

Changing e.g. the subscription key shouldn't cause a render.

I noticed a lot of extra renders happening in my components with
useLiveQuery which I traced to this problem.