perf: Change useState to useRef/useMemo for values not used for rendering (#678)
This is the general rule for state vs. ref — only use state if it's
rendered — which in the case of useLiveQuery is just the results.

Changing e.g. the subscription key shouldn't cause a render.

I noticed a lot of extra renders happening in my components with
useLiveQuery which I traced to this problem.

Without this PR, my component rendered 8 times, with it,
it renders 2 times (starting the query and then showing results).