chore(electric): Cleanups in MigrationConsumer; ignore incoming Relation messages (#680)
I have extracted this from my work on VAX-1040.

Per Postgres' logical replication protocol, the publisher must send a
Relation message to the subscriber before sending the first
insert/update/delete operation for a given table since the current
connection was established. In Electric, we don't have a use for those
message since we maintain information about the electrified schema up to
date outside of the replication connection.

I've also snuck in an optimization for the count of electrified tables
that we use in a telemetry event.