feat (DAL): client-side support for float4 (#671)
This PR adds client-side support for float4.
Numbers provided by the user are rounded to the nearest float4
representation using `Math.fround()`.
Similarly, when numbers are read from a float4 column in the DB they are
rounded to the nearest float4 representation.
This would not strictly be necessary if all writes happen through the
DAL but because of SQLite's dynamic typing, users can write
larger/smaller numbers to float4 columns in which case we would read
numbers that fall out of range.