feat(DAL): client-side support for JSON type (#633)
This PR adds client-side support for the JSON type.
The DAL accepts JS values representing JSON and serialises them to a
string that is stored in SQLite.
When reading a JSON value from the database the string is deserialised
back into a JS value.
A corner case arises when having an optional column of type JSON because
we need to differentiate between the database NULL value and the JSON
null value. We treat the regular JS null value as a database NULL (to be
consistent with how null values are interpreted for other column types)
and require users to pass a special JsonNull object in order to store a
top-level JSON null value.

Still need to add an E2E test for json values.

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>