fix (client): fix race condition in process.subscribe that made the client crash (#702)
This PR fixes the race condition described in #426.
The problem was that `process.subscribe` was creating a promise, storing
it, making an async call, fetching the promise and returning it. But,
depending on the order in which messages are received from Electric, the
async call may resolve after the subscription was already delivered and
hence the promise had already been deleted. The fix consists of storing
the promise in a local variable inside `process.subscribe` such that it
doesn't need to fetch the promise after the async call.