chore (client): expose a close method on the ElectricNamespace to dispose its resources (#703)
This PR addresses the first task of #243.
As suggested, the `ElectricNamespace` (and thus also the
`ElectricClient` as it extends the namespace) now exposes a `close`
method that can be called by the user to dispose all resources used by
the Electric client. This `close` method calls the `stop` on the global
registry such that it also stops the client's satellite process.