chore (client): refactoring subscribe methods of the event notifier (#708)
This PR refactors the various subscribe methods provided by the event
notifier. For every listener that was being subscribed, the event
notifier would generate a random key, store the listener under that key,
and return that key. To unsubscribe, the user had to pass that key to
the corresponding unsubscribe method which would lookup the listener and
then unsubscribe it from the underlying event emitter. Storing the
listener under a key is not needed and is a bit redundant as the
underlying event emitter already stores them.

This PR modified the event notifier such that it does not keep track of
registered callbacks but instead returns an unsubscribe function from
the subscribe methods. The unsubscribe function closes over the callback
that needs to be unregistered from the underlying event emitter.