Make Electric.Extension.Case more flexible
Depending on whether proxy_opts are passed to it, we either
open a regular connection to the singleton test db (current behaviour
without proxy) or create a new temporary database and open a proxied
connection to it.

One of the reasons being the latest internal migration failing with a
SyntaxError when applied via the proxy.