Update Types doc to reference main branch (#745)
This link just threw me for a *huge* loop when I thought it meant I'd
have to abandon Electric after spending awhile setting it up, since it
references a significantly outdated version of this file. Not sure if
this is how you want to handle it though, or just moving up the commit
pointer.

Thanks for the great work, excited to try it out!