Remove the reference to supported_types() from the Types page
Without versioning support in our docs, we cannot maintain the proper alignment between
the source code and the published docs. Since we already keep the list of supported types
up-to-date with the latest published release of Electric, there's no need to include a link
to the source code on that page.