capacitor-sqlite runInTransaction now takes the lock. (#677)
[EDIT] Following @kevin-dp 's work on generic adapters, this PR has been
repurposed. The concurrency issue is now taken care of by implementing
BatchDatabaseAdapter.

## ORIGINAL PR DESCRIPTION
This PR fixes a concurrency issue with the capacitor-sqlite database
adapter. The mutex was previously only acquired in the `transaction`
method, causing concurrency conflicts when calling `runInTransaction`.

Comments have been added to other public methods of the adapter in order
to clarify the intended behaviour of each method. In effect, `run`,
`execute` and `query` have ambiguous meanings depending on the layer
(electric driver -> capacitor-sqlite -> native SQLite). It is therefore
important to make sure that the methods as implemented correspond to
what is expected of `electric`'s API.