Fallback to unencrypted connections when DATABASE_REQUIRE_SSL=false
This is an edge case that's not handled by epgsql.

Epgsql can handle the case where it tries to establish an SSL connection
but gets back a "no encryption" response from Postgres. It will keep the
connection open, though it will be unencrypted.

However, Fly Postgres seems to always close its end of the connection
if it's not configured to run in SSL mode. Epgsql does not try to
reconnect in this case, so we do this ourselves after removing the SSL
connection option.