Stricter parsing of DATABASE_URL
We override the value of "replication" anyway, so it doesn't make sense
to allow it to be passed in.

As for the sslmode, we do not support all the options that psql does.
Our rules are simpler: use SSL by default but allow falling back to
unencrypted connections by setting DATABASE_REQUIRE_SSL to false.