chore(electric): Flip the default value of `DATABASE_REQUIRE_SSL` to `true` (#767)
This is a better default for any environment other than, perhaps, dev.

Until now, connecting to a Postgres that was configured to only accept
SSL connection would result in a cryptic
`invalid_authorization_specification` error (FWIW the same error may be
logged when Postgres is configured to only accept unencrypted
connection, but this is a rare edge case). That was the case with
Crunchy Bridge and [AWS
RDS](https://github.com/electric-sql/electric/issues/543#issuecomment-1801897232).

With this new default, more hosted databases should be supported out of
the box. One prominent exception is Fly Postgres (at least prior to its
takeover by Supabase): since Fly machine all connect to the same
encrypted private network, there is no benefit of an additional
encryption for database connections on top of that. For cases like this,
we need to print a friendly error message instructing the developer to
set `DATABASE_REQUIRE_SSL=false` in their Electric configuration. This
will be addressed in https://github.com/electric-sql/electric/pull/762.

Note that Electric also needs to be able to verify the database host's
authenticity to prevent man-in-the-middle attacks. I've filed VAX-1446
to address this at a later time.