chore(electric): Clean up templated SQL routines (#511)
This is a cleanup PR that includes the following changes:
* Remove some of the dead SQL code from Extension migration modules. The
code used to define routines but they are now defined at run time.
* Import the Extension module into Functions to make all functions from
the former available automatically to all templated SQL routines. This
removes the need to maintain a map of assigns by hand.
* Fix the wildcard pattern in Functions to include subdirectories. Prior
to this change, the functions from
`lib/electric/postgres/extension/functions/ddlx/` weren't included in
the module.