feature(client): New client CLI commands (#664)
New CLI commands built into the client.

To run the CLI in a project that has electric-sql installed:
```
npx electric-sql
```
This will output the possible commands.

Run `npx electric-sql help [command-name]` for details of how to run a
command.

All config args can be loaded as env vars, including from a .env at the
root of the project.