fix(client_cli): Proxy fixes  (#825)
A couple of minor issues I found:

1. ~~Use the prisma user in the db proxy url, otherwise non elecrified
tables are generated.~~ Bad call, the prisma user is correctly being
used for the introspection. The proxy url doesn't always need to use the
prisma user. We can however log the correct proxy url used when
generating the client, because the one in the config differs the one
used for the introspection.

https://github.com/electric-sql/electric/blob/7fbf292ef1bcb07a71fd4e2a528d2cda965ac5ce/clients/typescript/src/cli/migrations/migrate.ts#L73

https://github.com/electric-sql/electric/blob/7fbf292ef1bcb07a71fd4e2a528d2cda965ac5ce/clients/typescript/src/cli/migrations/migrate.ts#L303-L309

2. Handle ssl mode when building the db url, so that when running
locally, the db and proxy urls are built with `?sslmode=disable`. Some
postgres migration clients won't connect without that flag in the url
when running without ssl.

---------

Co-authored-by: Sam Willis <sam.willis@gmail.com>