fix (client): use WebSocketWeb as the default socket for capacitor (#830)
This PR fixes the default socket that is used by the capacitor driver to
be the `WebSocketWeb` socket.
Note that the cordova driver is still using `MockSocket` as the default.
Not sure what should be the default for the Cordova driver? And whether
it's worth spending any time in this since it is not used.