fix (client): clear all subscriptions from the DB in 1 transaction (#880)
When a client reconnects to Electric and can't resume it resets its
local state (i.e. unsubscribes all subscriptions) and starts from
scratch. Currently, we unsubscribe subscription per subscription. For
each subscription, we delete the contents of the shapes from the DB.
However, this can lead to FK violations. e.g.:

- Table User
- Table Post has a FK to user
- Subscription 1 subscribes to user table
- Subscription 2 subscribes to post and user table

When clearing all subscriptions, if we first unsubscribe from
subscription 1, we are deleting the contents of the user table and
thereby violating any FKs that are in the post table.

This PR avoids this problem by clearing all subscriptions within a
single database transaction that defers the foreign keys until the end
of the transaction. Also introduced a unit test that reproduces the FK
violation and is gone with this fix.